py/pyproject.toml
[build-system]
requires = ["hatchling>=1.21"]
build-backend = "hatchling.build"

[project]
name = "seebiang"
version = "0.1.0"
description = "Static and animated visualisations of the Biáng character and its variants."
readme = "../README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "See Biáng! contributors" }]
dependencies = [
    "cairosvg>=2.7",
    "pillow>=10.0",
    "imageio>=2.34",
    "jsonschema>=4.22",
    "svgpathtools>=1.6",
    "fonttools>=4.50",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pixelmatch>=0.3",
    "playwright>=1.44",
]

[project.scripts]
seebiang = "seebiang.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["seebiang"]
